V.01.07 Facturación al 80%, procesos al 60%, permisosos al 40%

+++ app/Http/Livewire/General/DeleteModel.php
@@ -3,22 +3,25 @@
 namespace App\Http\Livewire\General;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Illuminate\Support\Facades\DB;
 use Livewire\Component;
 
 class DeleteModel extends Component
 {
-
+    use AuthorizesRequests;
     public $title,  $permission;
-        public $model, $event="";
+    public $model, $event = "";
     public function render()
     {
         return view('livewire.general.delete-model');
     }
     public function deleteModel()
     {
-        auth()->user()->hasPermissionTo($this->permission);
+        $this->authorize($this->permission);
+        
         $this->model->delete();
-        $this->emit('showAlert',"{$this->title} eliminado", 'success');
+        $this->emit('showAlert', "{$this->title} eliminado", 'success');
         $this->emit($this->event);
     }
 }
